+++ rectApp/src/main/java/rectApp/Point.java
@@ -2,15 +2,15 @@ package rectApp;
 
 /**Contains the field for points x and y */
 public class Point {
-    public int x;
-    public int y;
+    public float x;
+    public float y;
 
     /**
      * A point constructor
      * @param x
      * @param y
      */
-    Point(int x, int y){
+    Point(float x, float y){
         this.x = x;
         this.y = y;
     }
